refactor(call): extract PlatformBridge mixin from CallBloc (#1017)
Phase 2 of CallBloc decomposition — moves all CallkeepDelegate method
implementations into a _PlatformBridgeMixin defined in platform_bridge.dart.

The mixin lives as a part of call_bloc.dart (same library) so it retains
access to private event types. CallBloc gains the mixin via `with` and
loses the explicit `implements CallkeepDelegate`.

Moves ~180 lines out of call_bloc.dart (2879 → 2699):
- continueStartCallIntent + _continueStartCallIntent
- didPushIncomingCall
- performStartCall/Answer/End/SetHeld/SetMuted/SendDTMF/AudioDevice*
- didActivateAudioSession / didDeactivateAudioSession / didReset
- _perform() helper
